Now compiles
svn path=/branches/work/soprano2/; revision=698539

+++ soprano/query/query.cpp
@@ -20,7 +20,7 @@ float Soprano::Query::Float::value() const
     return m_value;
 }
 
-Soprano::Query::Double::Double( int value )
+Soprano::Query::Double::Double( double value )
     :m_value(value)
 {
 }
